czech.po additional translations (43% completed)
As I said before, i will be happy when may test my translations in-game. Nothing is perfect, male / female translations can be the same because I dont known what dialogue is for what sex. Some functions canot be tested without creator mode and some of that may be wrong  too.